change of data type for price

+++ category/models.py
@@ -5,9 +5,7 @@ class Category(models.Model):
     description = models.TextField(max_length=255, verbose_name=('Descripción'))
     is_active = models.BooleanField(default=True, verbose_name='Activo')
     is_moderated = models.BooleanField(default=True, verbose_name='Moderado')
-    price = models.DecimalField(
-        max_digits=100,
-        decimal_places=2,
+    price = models.PositiveIntegerField(
         null=True,        # Permite valores nulos en la base de datos
         blank=True,       # Permite que el campo esté vacío en formularios
         verbose_name=('Costo')
